% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.logistic_copula}
\alias{predict.logistic_copula}
\title{predict.logistic_copula}
\usage{
\method{predict}{logistic_copula}(object, new_x, ...)
}
\arguments{
\item{object}{The model object as returned by fit_copula_interactions}

\item{new_x}{A matrix of covariate values to compute predictions for.}
}
\description{
Computes predicted probability of Y=1 for a logistic regression model with
a vine extension.
}
