% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(
  y,
  x,
  m_obj,
  maxit = 5,
  num_grad = FALSE,
  verbose = FALSE,
  hessian = FALSE,
  reltol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{y}{A vector of n observations of the (univariate) binary outcome
variable y}

\item{x}{A (n x p) matrix of n observations of p covariates}

\item{m_obj}{The model object as returned from fit_copula_interactions}

\item{maxit}{The maximum number of gradient steps}

\item{verbose}{Whether information about the progress should be printed 
to the console.}

\item{hessian}{Whether to numerically compute the hessian matrix, see the
documentation for \link[stats]{optim}.}

\item{reltol}{Relative convergence tolerance, see the documentation for 
\link[stats]{optim}.}

\item{xtype}{A vector of p characters that have to take the value}

\item{num_grd}{Whether to compute gradients numerically.}
}
\description{
This function updates the parameters of a LogisticCopula 
model by maximum likelihood.
}
